requirements.txt
absl-py==1.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
deeplite-profiler==1.1.12
fonttools==4.33.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.3
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.2
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
